Tfl_Tube_Json

Documentation

firstlevel :: (t -> [Char]) -> [t] -> [Char]

The firstlevel function take the parameter fstr and a list of Json strings and gets the strings corespoding to the parameter fstr. This is used for the main Json text.

secondlevel :: (t -> a) -> [t] -> [a]

The secondlevel function take the parameter sstr and a list of Json strings and gets the strings corespoding to the parameter sstr. this is used for Json text that is nested in the main text

firstLevelNum :: (t -> a) -> [t] -> [a]

The firstLevelNum function take the parameter nstr and a list of Json intergers and gets the integers corespoding to the parameter nstr. This is used for the main Json integer types.

convToMin :: Integral a => a -> a

The convToMin function converts a given Int value to the lowest minute value

listConvert :: (Char -> Bool) -> String -> [String]

The listConvert function converts a given String str and converts it to a list of String arrays

statusConversion :: [Lines] -> IO Bool

The statusConversion function zips given Strings for use in storeStatus.

napId_Conversion :: [Stoppoint] -> IO Bool

The napId_Conversion function zips given Strings for use in storeNaptan.

arrivals_Conversion :: [Arrivals] -> IO Bool

The arrivals_Conversion function zips given Strings for use in storeArrivals.

getJSON_Status :: IO ByteString

getJSON_NaptanId :: IO ByteString

The getJSON_NaptanId get Json data from the given naptanId retrival Url.

retrieve_status :: IO (Either String [Lines])

The retrieve_status parses Json data from the given tube status Url.

retrieve_NaptanId :: IO (Either String [Stoppoint])

The retrieve_NaptanId parses Json data from the given naptanId retrival Url.

retrieveLineStatus :: IO ()

The retrieveLineStatus function retrieves the line status and outputs the parsed data.

retrieveNaptanId_with_Arrivals :: IO ()

The retrieveLineStatus function retrieves all naptan ids then get the station name from the user using request inputs it to retrieveArrivals to get arrival data for user specified station.

retrieveArrivals :: String -> IO ()

The retrieveArrivals function gets that arrival data for a specified station nap_Id.